test: add tests for group service handler and repository

+++ internal/api/v1beta1/group.go
@@ -6,6 +6,7 @@ import (
 	"github.com/raystack/frontier/core/audit"
 
 	"github.com/raystack/frontier/pkg/str"
+	"github.com/raystack/frontier/pkg/utils"
 
 	"github.com/pkg/errors"
 	"github.com/raystack/frontier/pkg/metadata"
@@ -69,9 +70,14 @@ func (h Handler) ListGroups(ctx context.Context, request *frontierv1beta1.ListGr
 func (h Handler) ListOrganizationGroups(ctx context.Context, request *frontierv1beta1.ListOrganizationGroupsRequest) (*frontierv1beta1.ListOrganizationGroupsResponse, error) {
 	logger := grpczap.Extract(ctx)
 
+	orgId := request.GetOrgId()
+	if !utils.IsValidUUID(orgId) {
+		return nil, grpcOrgNotFoundErr
+	}
+
 	var groups []*frontierv1beta1.Group
 	groupList, err := h.groupService.List(ctx, group.Filter{
-		OrganizationID: request.GetOrgId(),
+		OrganizationID: orgId,
 		State:          group.State(request.GetState()),
 	})
 	if err != nil {
@@ -227,7 +233,16 @@ func (h Handler) UpdateGroup(ctx context.Context, request *frontierv1beta1.Updat
 func (h Handler) ListGroupUsers(ctx context.Context, request *frontierv1beta1.ListGroupUsersRequest) (*frontierv1beta1.ListGroupUsersResponse, error) {
 	logger := grpczap.Extract(ctx)
 
-	users, err := h.userService.ListByGroup(ctx, request.Id, group.MemberPermission)
+	if !utils.IsValidUUID(request.GetOrgId()) {
+		return nil, grpcOrgNotFoundErr
+	}
+
+	id := request.GetId()
+	if !utils.IsValidUUID(id) {
+		return nil, grpcGroupNotFoundErr
+	}
+
+	users, err := h.userService.ListByGroup(ctx, id, group.MemberPermission)
 	if err != nil {
 		logger.Error(err.Error())
 		return nil, grpcInternalServerError
@@ -250,45 +265,108 @@ func (h Handler) ListGroupUsers(ctx context.Context, request *frontierv1beta1.Li
 
 func (h Handler) AddGroupUsers(ctx context.Context, request *frontierv1beta1.AddGroupUsersRequest) (*frontierv1beta1.AddGroupUsersResponse, error) {
 	logger := grpczap.Extract(ctx)
-	if err := h.groupService.AddUsers(ctx, request.GetId(), request.GetUserIds()); err != nil {
+
+	if !utils.IsValidUUID(request.GetOrgId()) {
+		return nil, grpcOrgNotFoundErr
+	}
+
+	id := request.GetId()
+	if !utils.IsValidUUID(id) {
+		return nil, grpcGroupNotFoundErr
+	}
+
+	if err := h.groupService.AddUsers(ctx, id, request.GetUserIds()); err != nil {
 		logger.Error(err.Error())
-		return nil, status.Errorf(codes.Internal, err.Error())
+		return nil, grpcInternalServerError
 	}
 	return &frontierv1beta1.AddGroupUsersResponse{}, nil
 }
 
 func (h Handler) RemoveGroupUser(ctx context.Context, request *frontierv1beta1.RemoveGroupUserRequest) (*frontierv1beta1.RemoveGroupUserResponse, error) {
 	logger := grpczap.Extract(ctx)
+	if !utils.IsValidUUID(request.GetOrgId()) {
+		return nil, grpcOrgNotFoundErr
+	}
+
+	id := request.GetId()
+	if !utils.IsValidUUID(id) {
+		return nil, grpcGroupNotFoundErr
+	}
 	if err := h.groupService.RemoveUsers(ctx, request.GetId(), []string{request.GetUserId()}); err != nil {
 		logger.Error(err.Error())
-		return nil, status.Errorf(codes.Internal, err.Error())
+		return nil, grpcInternalServerError
 	}
 	return &frontierv1beta1.RemoveGroupUserResponse{}, nil
 }
 
 func (h Handler) EnableGroup(ctx context.Context, request *frontierv1beta1.EnableGroupRequest) (*frontierv1beta1.EnableGroupResponse, error) {
 	logger := grpczap.Extract(ctx)
-	if err := h.groupService.Enable(ctx, request.GetId()); err != nil {
+
+	if !utils.IsValidUUID(request.GetOrgId()) {
+		return nil, grpcOrgNotFoundErr
+	}
+
+	id := request.GetId()
+	if !utils.IsValidUUID(id) {
+		return nil, grpcGroupNotFoundErr
+	}
+
+	if err := h.groupService.Enable(ctx, id); err != nil {
 		logger.Error(err.Error())
-		return nil, status.Errorf(codes.Internal, err.Error())
+		switch {
+		case errors.Is(err, group.ErrNotExist):
+			return nil, grpcGroupNotFoundErr
+		default:
+			return nil, grpcInternalServerError
+		}
 	}
 	return &frontierv1beta1.EnableGroupResponse{}, nil
 }
 
 func (h Handler) DisableGroup(ctx context.Context, request *frontierv1beta1.DisableGroupRequest) (*frontierv1beta1.DisableGroupResponse, error) {
 	logger := grpczap.Extract(ctx)
-	if err := h.groupService.Disable(ctx, request.GetId()); err != nil {
+
+	if !utils.IsValidUUID(request.GetOrgId()) {
+		return nil, grpcOrgNotFoundErr
+	}
+
+	id := request.GetId()
+	if !utils.IsValidUUID(id) {
+		return nil, grpcGroupNotFoundErr
+	}
+
+	if err := h.groupService.Disable(ctx, id); err != nil {
 		logger.Error(err.Error())
-		return nil, status.Errorf(codes.Internal, err.Error())
+		switch {
+		case errors.Is(err, group.ErrNotExist):
+			return nil, grpcGroupNotFoundErr
+		default:
+			return nil, grpcInternalServerError
+		}
 	}
 	return &frontierv1beta1.DisableGroupResponse{}, nil
 }
 
 func (h Handler) DeleteGroup(ctx context.Context, request *frontierv1beta1.DeleteGroupRequest) (*frontierv1beta1.DeleteGroupResponse, error) {
 	logger := grpczap.Extract(ctx)
-	if err := h.groupService.Delete(ctx, request.GetId()); err != nil {
+
+	if !utils.IsValidUUID(request.GetOrgId()) {
+		return nil, grpcOrgNotFoundErr
+	}
+
+	id := request.GetId()
+	if !utils.IsValidUUID(id) {
+		return nil, grpcGroupNotFoundErr
+	}
+
+	if err := h.groupService.Delete(ctx, id); err != nil {
 		logger.Error(err.Error())
-		return nil, status.Errorf(codes.Internal, err.Error())
+		switch {
+		case errors.Is(err, group.ErrNotExist):
+			return nil, grpcGroupNotFoundErr
+		default:
+			return nil, grpcInternalServerError
+		}
 	}
 	return &frontierv1beta1.DeleteGroupResponse{}, nil
 }
